Pequena correcao.

+++ software/pc.c.h
@@ -12,7 +12,7 @@ void PC_Restart (void) {
     Serial.print(F(";"));
     Serial.print(S.timeout/1000);       // tempo total de jogo
     Serial.print(F(";"));
-    Serial.print(S.equilibrio);         // equilibrio ligado?
+    Serial.print((int)S.equilibrio);    // equilibrio ligado?
     Serial.print(F(";"));
     Serial.print(S.names[0]);           // atleta a esquerda
     Serial.print(F(";"));
